requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
requests==2.32.4
python-dotenv==1.1.1
psutil==7.0.0
scapy==2.6.1
yara-python==4.3.2
shodan==1.31.0
dnspython==2.7.0
geoip2==5.1.0
netifaces==0.11.0
Flask-SocketIO==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2
plotly==6.2.0
dash==3.1.0
pandas==2.3.0
numpy==2.0.2
rich==14.0.0
colorama==0.4.6
click==8.1.8
prompt_toolkit==3.0.51
pygments==2.19.2
# New dependencies for enhanced features
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==5.1.0
